update video comp editor

+++ pychron/core/ui/qt/video_component_editor.py
@@ -47,7 +47,7 @@ class _VideoComponentEditor(_LaserComponentEditor):
         # self.control.connect(self.playTimer, SIGNAL('timeout()'), self.update)
         print("fffffffffff", self.value.fps)
         if self.value.fps:
-            self.playTimer.setInterval(1000 / float(self.value.fps))
+            self.playTimer.setInterval(int(1000 / float(self.value.fps)))
         self.playTimer.start()
 
         self.value.on_trait_change(self.stop, "closed_event")
